feat: setup theme package for base styles
