+++ package.json
@@ -39,7 +39,8 @@
     "storybook-addon-performance": "^0.16.1"
   },
   "scripts": {
-    "build": "yarn workspace @pretear/select run build",
+    "select:build": "yarn workspace @pretear/select run build",
+    "theme:build": "yarn workspace @pretear/theme run build",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook"
   }
