+++ packages/select/src/index.ts
@@ -1,1 +1,1 @@
-export * from './Select';
+export * from './select';
