+++ packages/theme/src/index.ts
@@ -1,0 +1,2 @@
+import * as theme from './theme';
+export default theme;
