+++ packages/theme/src/theme/components/index.ts
@@ -1,0 +1,1 @@
+export { default as Select } from './select';
